requirements.txt
alembic==1.8.1
click==8.1.3
Flask==2.2.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
Mako==1.2.4
MarkupSafe==2.1.1
Pillow==9.3.0
python-pptx==0.6.21
SQLAlchemy==1.4.44
Werkzeug==2.2.2
WTForms==3.0.1
XlsxWriter==3.0.3
